feat(config): support creating config via `--write-config`
